Add memtrace support
