+++ src/core.param.def
@@ -286,5 +286,15 @@ DEF_PARAM(cbp_trace_r61, CBP_TRACE_R61, char*, string, NULL, )
 DEF_PARAM(cbp_trace_r62, CBP_TRACE_R62, char*, string, NULL, )
 DEF_PARAM(cbp_trace_r63, CBP_TRACE_R63, char*, string, NULL, )
 
+DEF_PARAM(memtrace_modules_log, MEMTRACE_MODULES_LOG, char*, string, NULL, )
+DEF_PARAM(prefetcher_prio_list, PREFETCHER_PRIO_LIST, char*, string, "", )
+DEF_PARAM(pref_prio_on, PREF_PRIO_ON, uns, uns, 0, )
+DEF_PARAM(pref_prio_add_branches, PREF_PRIO_ADD_BRANCHES, uns, uns, 0, )
+DEF_PARAM(pref_prio_dram, PREF_PRIO_DRAM, uns, uns, 0, )
+DEF_PARAM(horob_min_cycles, HOROB_MIN_CYCLES, uns, uns, 0, )
+DEF_PARAM(miss_count_path, MISS_COUNT_PATH, char*, string, NULL, )
+DEF_PARAM(pref_learn_deps, PREF_LEARN_DEPS, uns, uns, 0, )
+DEF_PARAM(pref_min_miss_ratio, PREF_MIN_MISS_RATIO, uns, uns, 10, )
+
 DEF_PARAM(dumb_core_on, DUMB_CORE_ON, Flag, Flag, FALSE, )
 DEF_PARAM(dumb_core, DUMB_CORE, uns, uns, 1, )
