+++ src/pin/pin_lib/Makefile
@@ -9,13 +9,13 @@ VERSION = 0
 
 SCARAB_DIR ?= ../../
 OBJDIR := obj
-CCFILES = message_queue_interface_lib.cc pin_scarab_common_lib.cc
+CCFILES = message_queue_interface_lib.cc pin_scarab_common_lib.cc x86_decoder.cc x87_stack_delta.cc
 CFILES = uop_generator.c
 OBJFILES = $(patsubst %.cc,$(OBJDIR)/%.o,$(CCFILES)) $(patsubst %.c,$(OBJDIR)/%.o,$(CFILES))
 TARGET = libpin_fe
 
 DEBUG_FLAGS = -DNO_DEBUG
-XFLAGS	    = -D_REENTRANT
+XFLAGS	    = -D_REENTRANT -DMEMTRACE
 WARN_CFLAGS = -Wall -Wunused -Wmissing-prototypes -Wmissing-declarations -Wno-long-long -Wpointer-arith -Wimplicit -Werror
 ifeq ($(CCVER), gcc (GCC) 4.7.3)
 WARN_CFLAGS += -Wno-unused-but-set-variable -Wno-maybe-uninitialized
@@ -24,9 +24,9 @@ WARN_CCFLAGS= -Wall -Wunused -Wno-long-long -Wpointer-arith -Werror
 
 NO_STAT ?=
 PIN_COMPILE ?=
-INCLUDES = -I$(SCARAB_DIR)
-OPTCFLAGS   = -std=gnu99 -O3 $(ARCHTYPE) -funroll-loops -Werror $(INCLUDES) $(DEBUG_FLAGS) $(WARN_CFLAGS) $(XFLAGS) $(NO_STAT) $(PIN_COMPILE)# -DNO_ASSERT 
-OPTCCFLAGS   = -std=c++14 -O3 $(ARCHTYPE) -fno-inline -funroll-loops -Werror $(INCLUDES) $(DEBUG_FLAGS) $(WARN_CCFLAGS) $(XFLAGS) $(NO_STAT) $(PIN_COMPILE)# -DNO_ASSERT 
+INCLUDES = -I$(SCARAB_DIR) -I$(PIN_ROOT)/
+OPTCFLAGS   = -std=gnu99 -O3 -g $(ARCHTYPE) -funroll-loops -Werror $(INCLUDES) $(DEBUG_FLAGS) $(WARN_CFLAGS) $(XFLAGS) $(NO_STAT) $(PIN_COMPILE)
+OPTCCFLAGS   = -std=c++14 -O3 -g $(ARCHTYPE) -fno-inline -funroll-loops -Werror $(INCLUDES) $(DEBUG_FLAGS) $(WARN_CCFLAGS) $(XFLAGS) $(NO_STAT) $(PIN_COMPILE)
 
 DEPDIR := $(OBJDIR)
 DEPFLAGS = -MT $@ -MMD -MP -MF $(DEPDIR)/$*.d
